pyproject.toml
[project]
name = "codos"
version = "0.1.0"
description = "Atlas AI Operating System - unified service dependencies"
requires-python = ">=3.13"
dependencies = [
    "annotated-doc>=0.0.4",
    "annotated-types>=0.7.0",
    "anthropic>=0.77.0",
    "anyio>=4.12.1",
    "assemblyai>=0.49.0",
    "certifi>=2026.1.4",
    "charset-normalizer>=3.4.4",
    "click>=8.3.1",
    "composio>=0.11.0",
    "composio-client>=1.27.0",
    "distro>=1.9.0",
    "docstring-parser>=0.17.0",
    "fastapi>=0.128.0",
    "h11>=0.16.0",
    "httpcore>=1.0.9",
    "httptools>=0.7.1",
    "httpx>=0.28.1",
    "idna>=3.11",
    "jiter>=0.12.0",
    "json-schema-to-pydantic>=0.4.8",
    "linkify-it-py>=2.0.3",
    "markdown-it-py>=4.0.0",
    "mcp[cli]>=1.8.0",
    "mdit-py-plugins>=0.5.0",
    "mdurl>=0.1.2",
    "nest-asyncio>=1.6.0",
    "openai>=2.16.0",
    "pillow>=12.1.0",
    "platformdirs>=4.5.1",
    "pyaes>=1.6.1",
    "pyasn1>=0.6.2",
    "pydantic>=2.12.5",
    "pydantic-core>=2.41.5",
    "pydantic-settings>=2.0.0",
    "pygments>=2.19.2",
    "pysher>=1.0.8",
    "python-dotenv>=1.2.1",
    "loguru>=0.7.0",
    "python-multipart>=0.0.22",
    "python-telegram-bot>=21.0",
    "pyyaml>=6.0.3",
    "qrcode>=8.2",
    "requests>=2.32.5",
    "rich>=14.3.1",
    "rsa>=4.9.1",
    "sniffio>=1.3.1",
    "starlette>=0.50.0",
    "telethon>=1.42.0",
    "textual>=7.4.0",
    "tqdm>=4.67.1",
    "typing-inspection>=0.4.2",
    "typing-extensions>=4.15.0",
    "uc-micro-py>=1.0.3",
    "urllib3>=2.6.3",
    "uvicorn>=0.40.0",
    "uvloop>=0.22.1",
    "watchfiles>=1.1.1",
    "websocket-client>=1.9.0",
    "websockets>=16.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
backend = "backend.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/backend"]

[tool.pytest.ini_options]
testpaths = ["src/backend"]
asyncio_mode = "auto"

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.ruff]
line-length = 120
target-version = "py313"
exclude = [
    ".venv",
    "node_modules",
    "dev/desktop/src-tauri/target",
    "dev/desktop/src-tauri/resources",
    "dist",
    "build",
    "__pycache__",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["codos"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
